refactor: product/display 디렉토리 상위폴더로 통합
